Add support for -uninstalled pkg-config files.
By creating an -uninstalled variant of the pkg-config file that is not
installed but just left in the build directory structure, it is
possible to link to a live version of libxr without having to install
it anywhere in the system.

This makes it useful to build dependent packages all in a build tree
and test them right away, insulated from the system packages.